Skriv om testfil til TypeScript slik at vi kan sikre at datamodellane i den held seg oppdatert ved endringar
Legg også til gjenbrukbare objekt for "ingen valgte filter", slik at ein berre treng oppdatere éin stad når modellen endrar seg.
